add todo to delete methods

+++ src/main/java/com/orderandwarehouse/app/controller/ComponentController.java
@@ -43,11 +43,15 @@ public class ComponentController {
 
     @DeleteMapping("/{id}")
     public HttpStatus delete(@PathVariable Long id) {
-        service.delete(id);
-        //todo check what happens if i try to delete one that is in use
-        //todo create exception handler, check if this returns NOT_FOUND or OK if exception happens
-        //todo return not_found if doesn't exist
-        return HttpStatus.OK;
+        /*
+        TODO check if ID exists? OK : NOT_FOUND
+                if ID exists check if it's in use? StorageUnitStillInUseException : OK
+        */
+        if (service.getById(id).isPresent()) {
+            service.delete(id);
+            return HttpStatus.OK;
+        }
+        return HttpStatus.NOT_FOUND;
     }
 
     @GetMapping(params = "nameLike")
